閲覧、編集統合
